restaurant_orders/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
alembic>=1.7.1
python-telegram-bot>=13.7
pydantic>=1.8.2
pydantic-settings>=2.0.0
python-dotenv>=0.19.0
psycopg2-binary>=2.9.1
langchain>=0.0.200
python-escpos>=3.0
websockets>=10.0
python-multipart>=0.0.5
aiofiles>=0.8.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
